moving method to get web info into SPSitesService
