+++ src/services/FilesSearchService.types.ts
@@ -68,9 +68,3 @@ export interface Thumbnail {
   width: number;
   height: number;
 }
-
-export interface ISiteWebInfo {
-  title: string,
-  webId: string,
-  siteId: string
-}
